app/requirements.txt
# Core UI and visualization
streamlit>=1.28.0
streamlit-option-menu>=0.4.0
plotly>=6.2.0
matplotlib>=3.10.6

# Data processing and analysis
numpy>=2.3.1
pandas>=2.3.1
pillow>=10.0.0

# Database and storage
chromadb>=0.4.14

# LangChain and AI components
langchain-chroma>=0.2.5
langchain-community>=0.3.27
langchain-openai>=0.3.28
langchain-huggingface>=0.3.1
langgraph>=0.6.3
langgraph-checkpoint-sqlite>=2.0.11
langgraph-supervisor==0.0.29
langsmith>=0.4.8
sentence-transformers>=2.2.2

# Utilities
loguru>=0.7.3
python-dotenv>=1.0.0
requests>=2.31.0
